Refactor everything

See merge request idatt2106_2022_02/boco-backend!40

+++ src/main/java/no/ntnu/idatt2106/model/DAO/CategoryDAO.java
@@ -23,7 +23,7 @@ public class CategoryDAO {
     public CategoryDAO() {
     }
 
-    public long getCategoryID() {
+    public int getCategoryID() {
         return this.categoryID;
     }
